+++ src/user.ts
@@ -1,0 +1,17 @@
+export class User {
+  name: string;
+  password: string;
+  connection: WebSocket;
+
+  constructor(name: string, password: string) {
+    this.name = name;
+    this.password = password;
+  }
+
+  send(data: string) {
+    this.connection.send(data);
+  }
+  updateConnection(connection: WebSocket) {
+    this.connection = connection;
+  }
+}
